alerta/mqtt_to_Alerta_service/requirements.txt
paho-mqtt>=1.6.0
alerta
pydantic>=2.0